requirements.txt
pandas
snowflake-snowpark-python==0.7.0
streamlit-aggrid
openpyxl